modularize code and make it clean
