--- src/cardStyles.css
@@ -1,104 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Poppins:ital,wght@0,100;0,200;0,400;0,500;1,200&family=Roboto:ital,wght@0,100;0,300;1,300;1,700&display=swap');
-
-:root {
-  --font-poppins: 'Poppins', sans-serif;
-  --font-roboto: 'Roboto', sans-serif;
-  --cards-shadow: rgba(0, 0, 0, 0.09) 0 2px 1px, rgba(0, 0, 0, 0.09) 0 4px 2px, rgba(0, 0, 0, 0.09) 0 8px 4px, rgba(0, 0, 0, 0.09) 0 16px 8px, rgba(0, 0, 0, 0.09) 0 32px 16px;
-}
-
-header {
-  border-radius: 6px;
-  box-shadow: var(--cards-shadow);
-  margin-bottom: 2px;
-}
-
-#atvs {
-  text-decoration: none;
-  font-weight: bold;
-}
-
-.main-sec {
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  grid-template-rows: repeat(33, 1fr);
-  grid-column-gap: 2px;
-  grid-row-gap: 2px;
-}
-
-.card {
-  background-color: rgba(182, 182, 182, 0.1);
-  border: 1px solid rgba(0, 0, 0, 0.1);
-  border-radius: 6px;
-  box-shadow: rgba(0, 0, 0, 0.1) 0 0 5px 0, rgba(0, 0, 0, 0.1) 0 0 1px 0;
-  padding: 6px;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
-.card-subtitle {
-  font-family: var(--font-poppins);
-  font-size: 0.9rem;
-  font-weight: bold;
-  color: #000;
-  display: flex;
-  align-items: center;
-  height: 2em;
-  width: 200px;
-  justify-content: space-between;
-}
-
-.likes {
-  display: flex;
-  align-items: center;
-  justify-self: flex-end;
-  height: 42px;
-  font-family: var(--font-roboto);
-}
-
-.comment-btn {
-  padding: 6px 12px;
-  text-align: center;
-  transition: 0.5s;
-  background-size: 200% auto;
-  color: black;
-  border-radius: 6px;
-  display: block;
-  border: 0;
-  font-family: var(--font-poppins);
-  font-weight: 600;
-  box-shadow: 0 0 14px -7px #f09819;
-  background-image: linear-gradient(45deg, #ff512f 0%, #f09819  51%, #ff512f  100%);
-  cursor: pointer;
-  user-select: none;
-
-  /* -webkit-user-select: none; */
-  touch-action: manipulation;
-}
-
-.comment-btn:hover {
-  background-position: right center;
-
-  /* change the direction of the change here */
-  color: #fff;
-  text-decoration: none;
-}
-
-.comment-btn:active {
-  transform: scale(0.95);
-}
-
-@media (max-width: 768px) and (min-width: 485px) {
-  .main-sec {
-    grid-template-columns: repeat(2, 1fr);
-    grid-template-rows: repeat(50, 1fr);
-  }
-}
-
-@media (max-width: 484px) {
-  .main-sec {
-    grid-template-columns: repeat(1, 1fr);
-    grid-template-rows: repeat(99, 1fr);
-  }
-}
