--- src/comment.css
@@ -1,52 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Poppins:ital,wght@0,100;0,200;0,400;0,500;1,200&family=Roboto:ital,wght@0,100;0,300;1,300;1,700&display=swap');
-
-:root {
-  --font-poppins: 'Poppins', sans-serif;
-  --font-roboto: 'Roboto', sans-serif;
-}
-
-.comment-section {
-  margin: 1rem 0;
-}
-
-.center-comment {
-  text-align: center;
-}
-
-.comment-list {
-  margin: 0 1rem;
-  font-family: var(--font-roboto);
-}
-
-.form-comment {
-  display: flex;
-  flex-direction: column;
-  gap: 0.75rem;
-  width: 50%;
-  margin: 0 auto;
-}
-
-.form-comment input {
-  padding: 0.5rem;
-  border-radius: 5px;
-  border: none;
-  outline: none;
-  font-family: var(--font-roboto);
-}
-
-.form-comment textarea {
-  padding: 0.5rem;
-  outline: none;
-  border: none;
-  border-radius: 5px;
-  color: #000;
-  font-family: var(--font-roboto);
-}
-
-.add-comment {
-  padding: 0.5rem;
-  align-self: center;
-  cursor: pointer;
-  border-radius: 5px;
-  font-family: var(--font-roboto);
-}
