+++ src/index.html
@@ -21,9 +21,9 @@
   <div class="modal hide">
   </div>
   <footer>
-    <h4>Copyright &copy; 2022 Created by Habeeb Efiamotu  and Lucas Mullen</h4>
-    <h5>All Rights Reserved</h5>
-    <h5>MIT Licensed</h5>  
+    <p>Copyright &copy; 2022 Created by Habeeb Efiamotu  and Lucas Mullen</p>
+    <p>All Rights Reserved</p>
+    <p>MIT Licensed</p>  
   </footer>
 </body>
 </html>
