--- src/style.css
@@ -1,49 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Poppins:ital,wght@0,100;0,200;0,400;0,500;1,200&family=Roboto:ital,wght@0,100;0,300;1,300;1,700&display=swap');
-
-:root {
-  --font-poppins: 'Poppins', sans-serif;
-  --font-roboto: 'Roboto', sans-serif;
-}
-
-body {
-  background-color: bisque;
-}
-
-h1 {
-  font-family: var(--font-poppins);
-}
-
-header nav ul {
-  list-style-type: none;
-  margin: 0;
-  padding: 0;
-  display: flex;
-  flex-direction: row;
-  font-family: var(--font-poppins);
-  font-size: 16px;
-  justify-content: space-around;
-  align-items: center;
-}
-
-#img-logo {
-  background-image: url(./imgs/logo.png);
-  background-repeat: no-repeat;
-  background-position: center;
-  background-size: cover;
-  height: 60px;
-  width: 120px;
-}
-
-footer {
-  background-color: bisque;
-  font-family: var(--font-poppins);
-  font-size: 16px;
-  padding: 20px;
-  text-align: center;
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  gap: 12%;
-  border: 2px solid;
-  align-items: center;
-}
